Fixed warnings for x86 version...

+++ MonsterChase/MonsterChase.cpp
@@ -197,7 +197,7 @@ void MonsterChase::ValidateName(const char* input)
 		count += isblank(c) ? 1 : 0;
 	}
 
-	size_t input_length = strlen(input);
+	int input_length = (int)strlen(input);
 	if (input_length > Player::MAX_NAME_LENGTH				// check if the name was within our range
 		|| count >= input_length - 1)						// check if the input contained only white spaces
 	{
